Add a license file
The dubious-looking "2010-2012" merits an explanation, so let's get this
pedantic bit of history cleared up: "Unitome" was a project I created in
the days before I used version control. None of the original files exist
in anything more than spirit: this project is essentially a continuation
of what I'd started and forgotten to finish - a character identification
tool which consults `UnicodeData.txt` directly. Though its original form
was that of a single-page HTML "webapp" loading a static text-file using
jQuery's `.load()` helper. More features were planned, but the project's
source-code got nuked whilst cleaning up projects I'd lost interest in.

I've no fucking idea how relevant an earlier incarnation of a project is
to a license file's attested year-range. But when each "Unitome" started
life as throwaway code that later formed the basis of a project with the
same vision, almost exactly 10 years apart... well, let's say the timing
was too perfect not to write a 72x15-size commit message documenting it.